add height margin to cutouts
when quick-rendered, the module looks aweful because everything is cut to
exact dimensions, so corners and edges are "hair-thin" in quick renders.
Only full renders will remove those infinitely small edges. So I add some
margin to cuts and change the base-cut from a copy of the outline to a
large cube. Also causes faster renders